classes for super computers
Its time I start a module that helps users build and run things on super computers.
The idea is a class that incompasses most user facing attributes.
Users can use this class and functions for this class to:
1)build on SCs
2)run on SCs
  a)interactive
  b)batch
3)quickly find results

We also want to create pickle file containing many different Super Computer objects.
We'll have to come up with some cool ideas about what are attributes of a Super Computer